+++ CHANGELOG.rst
@@ -1,3 +1,9 @@
+DEV
+------------------
+* Remove `with_ssl_support` flag for methods
+  `User.get_sso_create_site_url()` and `Site.change_site_domain()`.
+
+
 0.4.1
 ------------------
 * Add support of `with_ssl_support` flag for methods